requirements.txt
asgiref==3.2.9
dj-database-url==0.5.0
Django==3.0.7
django-heroku==0.3.1
djangorestframework==3.11.0
joblib==0.15.1
numpy==1.18.5
pandas==1.0.4
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
threadpoolctl==2.1.0
whitenoise==5.1.0
